Add: New stats and config updates for Python
The `ann-benchmarks` has an endpoint to estimate memory-usage
of some algorithm implementation. To support that natively,
instead of system calls, we add a `memory_usage()` method in C++
and a `memory_usage` computed property to Python.
For similar reasons, the `expansion_add` and `expansion_search`
are added as mutable computed properties.