Fix loading of Linux native lib from jar using NativeUtils.java from https://github.com/adamheinrich/native-utils.
NativeUtils works by copying the native lib to a temporary directory.

NOTE: as the current jar is not cross-platform and only includes the Linux native lib (libusearch.so), this is the native lib that is copied. When the native libs for other platforms are added, the code in Index.java needs to be updated to become platform-specific.