Use NSError instead of NSException
NSException indicates an unrecoverable runtime exception, and as a result cannot be caught in Swift. While it's possible to catch in Objective-C, it's still less ergonomic than the typical (NSError**) argument.

This change moves the Objective-C API to using the (NSError**) style errors, and a throwing Swift API with an error enum.

All methods that could reasonably throw have been converted, even though some do not currently throw, so that future changes, or errors surfaced from the C++ code, can be propagated without breaking changes.