Fix: `__usearch_raise_runtime_error` lexing UB (#622)
According to https://eel.is/c++draft/lex.name#3, symbols that
include double underscores and symbols in the global namespace
that start with an underscore are "reserved for use by C++
implementations and shall not be used otherwise".
As I understand the standards, that means it's technically UB
to use such symbols. This change renames the symbol to a legal
one.