javascript: improve typings
Add can take a matric due to normalizeVectors being called.

Additionally, support more strict typing in `search`: if you only search with a single vector input, you get a Matches object back. This simplifies the API instead of having to assert on the return type as a user of the library.