Improve: Support more input types in JS (#653)
`add` can now take a matrix due to `normalizeVectors` being called.

Additionally, support more strict typing in `search`: if you only search
with a single vector input, you get a `Matches` object back. This 
simplifies the API instead of having to assert on the return type as a
user of the library.

Co-authored-by: Tyler Rockwood <8669100+rockwotj@users.noreply.github.com>